通知系统

+++ web/pages/OrderManagement/OrderManagementController.js
@@ -10,7 +10,7 @@ angular.module("IntegratedFramework.OrderManagementController", ['ngRoute'])
         })
     }])
 
-    .controller('OrderManagementController', function ($scope, $http, myHttpService, serviceList, validate, notification, renderTableService, dispatchApsService) {
+    .controller('OrderManagementController', function ($scope, $http, myHttpService, serviceList, validate, notification, renderTableService, dispatchApsService, confirm) {
 
         layer.load(0);
 
@@ -61,19 +61,21 @@ angular.module("IntegratedFramework.OrderManagementController", ['ngRoute'])
         function confirmDispatchAps() {
             layer.load();
             setTimeout(function () {
-                layer.msg('已下发', {icon: 1});
+                notification.sendNotification("confirm", "已下发");
+                // layer.msg('已下发', {icon: 1});
                 hideLoadingPage();
             }, 2000);
         }
 
         //取消下发APS
         function resetDispatchAps() {
-            layer.msg('取消下发', {icon: 2});
+            notification.sendNotification("confirm", "取消下发");
+            // layer.msg('取消下发', {icon: 2});
         }
 
         //将选中记录下发APS
         $scope.dispatchRecord = function () {
-            dispatchApsService.dispatchAps(confirmDispatchAps,resetDispatchAps);
+            dispatchApsService.dispatchAps(confirmDispatchAps, resetDispatchAps);
         };
 
         //信息填写检验
@@ -137,6 +139,8 @@ angular.module("IntegratedFramework.OrderManagementController", ['ngRoute'])
                 return false;
             }
         };
+
+
 
 
         //信息填写检验
@@ -207,17 +211,19 @@ angular.module("IntegratedFramework.OrderManagementController", ['ngRoute'])
                 $("#modal-add").modal('hide');
                 console.log(addData);
                 myHttpService.post(serviceList.AddOrder, addData).then(function successCallback() {
-                    //用强制刷新解决按钮不能连续响应
-                    //location.reload(true);
-                    setTimeout('window.location.reload();', 0.1);
+                    myHttpService.get(serviceList.ListOrder).then(function (response) {
+                        $scope.orderList = response.data;
+                        notification.sendNotification("confirm", "添加成功");
+                    })
                 }, function errorCallback() {
                     notification.sendNotification("alert", "请求失败");
                 })
             } else {
-                notification.sendNotification("alert", "参数错误");
+                notification.sendNotification("alert", "输入有误");
             }
         };
 
+        //获得表单信息
         var getInfo = function () {
             $("div").removeClass("has-error");
             $("div").removeClass("has-success");
@@ -233,6 +239,7 @@ angular.module("IntegratedFramework.OrderManagementController", ['ngRoute'])
                 }
                 return true;
             } else {
+                notification.sendNotification("alert", "请重新选择！");
                 return false;
             }
         };
@@ -242,14 +249,12 @@ angular.module("IntegratedFramework.OrderManagementController", ['ngRoute'])
             if (getInfo()) {
                 $("#modal-edit").modal('show');
                 var idInfo = JSON.stringify(id_params);
-                console.log("&&&&&&&&&&&");
-                console.log(id_params);
                 myHttpService.post(serviceList.GetOrderById, idInfo).then(function successCallback(response) {
                     var editList = [];//保存从数据库获取的需要修改的数据
                     editList.push(response.data);
                     edit_params = response.data;
                     $scope.editList = editList;
-                }, function errorCallback(response) {
+                }, function errorCallback() {
                     notification.sendNotification("alert", "请求失败");
                 })
             } else {
@@ -258,23 +263,27 @@ angular.module("IntegratedFramework.OrderManagementController", ['ngRoute'])
         };
 
         $scope.editOrder = function () {
-            if (orderEditValidate()) {
-                $("#modal-edit").modal('hide');
-                //用获取到的数据代替从数据库取到的数据
-                edit_params.name = editData.name;
-                edit_params.quantity = editData.quantity;
-                edit_params.priority = editData.priority;
-                edit_params.t0 = editData.t0;
-                edit_params.t2 = editData.t2;
-                var update_data = angular.toJson(edit_params);
-                myHttpService.post(serviceList.UpdateOrder, update_data).then(function successCallback() {
-                    //location.reload(true);
-                    setTimeout('window.location.reload();', 0.1);
-                }, function errorCallback() {
-                    notification.sendNotification("alert", "请求失败");
-                })
-            } else {
-                notification.sendNotification("alert", "输入有误");
+            if (confirm.confirmEdit()) {
+                if (orderEditValidate()) {
+                    $("#modal-edit").modal('hide');
+                    //用获取到的数据代替从数据库取到的数据
+                    edit_params.name = editData.name;
+                    edit_params.quantity = editData.quantity;
+                    edit_params.priority = editData.priority;
+                    edit_params.t0 = editData.t0;
+                    edit_params.t2 = editData.t2;
+                    var update_data = angular.toJson(edit_params);
+                    myHttpService.post(serviceList.UpdateOrder, update_data).then(function successCallback() {
+                        myHttpService.get(serviceList.ListOrder).then(function (response) {
+                            $scope.orderList = response.data;
+                            notification.sendNotification("confirm", "修改成功");
+                        })
+                    }, function errorCallback() {
+                        notification.sendNotification("alert", "请求失败");
+                    })
+                } else {
+                    notification.sendNotification("alert", "输入有误");
+                }
             }
         };
 
@@ -282,12 +291,18 @@ angular.module("IntegratedFramework.OrderManagementController", ['ngRoute'])
         //删除订单
         $scope.deleteOrder = function () {
             if (getInfo()) {
-                var idInfo = JSON.stringify(id_params);
-                myHttpService.delete(serviceList.DeleteOrder, idInfo).then(function successCallback() {
-                    setTimeout('window.location.reload();', 0.1);
-                }, function errorCallback() {
-                    notification.sendNotification("alert", "请求失败");
-                });
+                if (confirm.confirmDel()) {
+                    var idInfo = JSON.stringify(id_params);
+                    myHttpService.delete(serviceList.DeleteOrder, idInfo).then(function successCallback() {
+                        // location.reload(true);
+                        myHttpService.get(serviceList.ListOrder).then(function (response) {
+                            $scope.orderList = response.data;
+                            notification.sendNotification("confirm", "删除成功");
+                        })
+                    }, function errorCallback() {
+                        notification.sendNotification("alert", "请求失败");
+                    });
+                }
             }
         };
 
@@ -296,49 +311,5 @@ angular.module("IntegratedFramework.OrderManagementController", ['ngRoute'])
             $("div").removeClass("has-error");
             $("div").removeClass("has-success");
         };
-
-        $scope.clicked = function (id) {
-            // $("#modal-info").modal('show');
-            o = document.getElementById(id).rows;//表格所有行
-            for (i = 0; i < o.length; i++) {
-                o[i].ondblclick = function () { //设置事件
-                    var oo;
-                    oo = this.cells[1]; //取得第二列对象
-                    alert(oo.innerHTML);
-                }
-            }
-        }
-
-        /*function clicked(id) {
-         var o, i;
-         o = document.getElementById(id).rows;//表格所有行
-         for (i = 0; i < o.length; i++) {
-         o[i].ondblclick = function () { //设置事件
-         var oo;
-         oo = this.cells[1]; //取得第二列对象
-         alert(oo.innerHTML);
-         }
-         }
-         }*/
-
-        window.onload = function () {
-            clicked('table_value'); //设置表格事件
-        }
-
-        /*$scope.getInfo = function () {
-         $("#modal-info").modal('show');
-         var trs = document.getElementsByTagName('tr');
-         for (var i = 1; i < trs.length; i++) {
-         var id = this.children[1].innerHTML;
-         console.log(id);
-         myHttpService.post(serviceList.GetOrderById, id).then(function successCallback(response) {
-         $scope.infoList = response.data;
-         }, function errorCallback() {
-         notification.sendNotification("alert", "请求失败");
-         })
-         break;
-         }
-         }*/
-
 
     });
